profile/blp/central/src/androidTest/java/org/im97mori/ble/profile/blp/central/db/BloodPressureProfileBondedDatabaseHelperTest.java
package org.im97mori.ble.profile.blp.central.db;

import androidx.test.core.app.ApplicationProvider;

import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;

public class BloodPressureProfileBondedDatabaseHelperTest {

    @Test
    public void test_constructor_00001() {
        assertNotNull(BloodPressureProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()));
    }

    @Test
    public void test_constructor_00002() {
        assertEquals(BloodPressureProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()), BloodPressureProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()));
    }

    @Test
    public void test_getProfileName_00001() {
        assertEquals(BloodPressureProfileBondedDatabaseHelper.PROFILE_NAME, BloodPressureProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()).getProfileName());
    }
}
